dsandasjkd

+++ results/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 class Result(models.Model):
     quiz = models.ForeignKey(Quiz, on_delete=models.CASCADE)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    score = models.FloatField()
-
+    score = models.IntegerField()
+    
     def __str__(self):
         return str(self.pk)
